Add test cases and query plotting

+++ worker_process.py
@@ -16,7 +16,7 @@ from worker_thread_fns import ( receive_images,send_messages,receive_messages,
                                consistency_function, write_data_csv)
 
 
-def worker(hosts,ports,audit_rate,worker_num, timeout = 20, VERBOSE = False):
+def worker(hosts,ports,audit_rate,worker_num, timeout = 20, VERBOSE = False,OVERWRITE = True):
     """
     Given a list of hosts and ports (one pair for each worker) and an id num,
     creates a series of threads and shared variables that carry out the work in
@@ -43,7 +43,7 @@ def worker(hosts,ports,audit_rate,worker_num, timeout = 20, VERBOSE = False):
     p_continue_consistency = mp.Value('b',True,lock = True)
     
     # changeable worker parameters
-    lb_timeout = 2
+    lb_timeout = 0.5
     query_timeout = 5
     consistency_rate = 0.5 # queries per second
     heart_rate = 0.05
@@ -60,10 +60,13 @@ def worker(hosts,ports,audit_rate,worker_num, timeout = 20, VERBOSE = False):
     ports.remove(port)
     
     # overwrite datafile and write one dummy line
-    data_path = os.path.join('databases','worker_{}_database.csv'.format(worker_num))
-    with open(data_path,mode = 'w') as f:
-        pass
-    write_data_csv(data_path,np.zeros([2,8])-1,-1,-1)
+    if OVERWRITE:
+        data_path = os.path.join('databases','worker_{}_database.csv'.format(worker_num))
+        p_database_lock.acquire()
+        with open(data_path,mode = 'w') as f:
+            pass
+        p_database_lock.release()
+        write_data_csv(data_path,np.zeros([2,8])-1,-1,p_database_lock,-1)
 
     
     # create image receiver thread
@@ -110,7 +113,7 @@ def worker(hosts,ports,audit_rate,worker_num, timeout = 20, VERBOSE = False):
                                      audit_rate,
                                      timeout,
                                      lb_timeout, 
-                                     VERBOSE,
+                                     True,
                                      len(ports)+1, # total num workers
                                      worker_num))
     
@@ -157,8 +160,7 @@ def worker(hosts,ports,audit_rate,worker_num, timeout = 20, VERBOSE = False):
                                      p_continue_consistency,
                                      consistency_rate, # queries per second
                                      query_timeout,
-                                     timeout,
-                                     True,
+                                     VERBOSE,
                                      worker_num))
                                     
     
@@ -181,29 +183,4 @@ def worker(hosts,ports,audit_rate,worker_num, timeout = 20, VERBOSE = False):
     t_heartbeat.join()
     t_query.join()
     t_consistency.join()
-    
-if __name__ == "__main__":
-    # tester code for the worker processes
-    
-    hosts = []
-    ports = []
-    num_workers = 4
-    timeout = 30
-    VERBOSE = False
-    audit_rate = mp.Value('f',0.1,lock = True)
-    
-    for i in range(num_workers):
-        hosts.append("127.0.0.1")
-        ports.append(5200+i)
-    
-    processes = []
-    for i in range(num_workers):
-        p = mp.Process(target = worker, args = (hosts,ports,audit_rate,i,timeout,VERBOSE,))
-        p.start()
-        processes.append(p)
-    print("All processes started")
-    
-    for p in processes:
-        p.join()
-    print("All processes terminated")
